requirements.txt
streamlit
pandas
pypdf
google-generativeai
pywebview
pyinstaller
